+++ src/SFML/Graphics/View.cpp
@@ -108,7 +108,7 @@ void View::setSize(const Vector2f& size)
 ////////////////////////////////////////////////////////////
 void View::setRotation(float angle)
 {
-    m_rotation = static_cast<float>(fmod(angle, 360));
+    m_rotation = std::fmod(angle, 360.f);
     if (m_rotation < 0)
         m_rotation += 360.f;
